+++ modules/counter/server-ts/__tests__/Counter.spec.ts
@@ -1,7 +1,7 @@
 import chai from 'chai';
 import { step } from 'mocha-steps';
 
-import { getServer, getApollo } from '../../../../packages/server-ts/src/testHelpers/integrationSetup';
+import { getServer, getApollo } from '../../../../packages/server/src/testHelpers/integrationSetup';
 
 import { COUNTER_QUERY, ADD_COUNTER, COUNTER_SUBSCRIPTION } from '@gqlapp/counter-common';
