+++ modules/counter/server-ts/sql.ts
@@ -1,4 +1,4 @@
-import knex from '../../../packages/server-ts/src/sql/connector';
+import knex from '../../../packages/server/src/sql/connector';
 
 export default class Counter {
   public counterQuery() {
