+++ packages/client/src/__tests__/entry.ts
@@ -1,6 +1,6 @@
 import { initRenderer } from '@gqlapp/testing-client-react';
 import rootSchema from '@gqlapp/core-server-ts/api/rootSchema.graphql';
 import clientModules from '..';
-import serverModules from '../../../server-ts/src/modules';
+import serverModules from '../../../server/src/modules';
 
 initRenderer([rootSchema].concat(serverModules.schema), clientModules);
